requirements.txt
pillow==3.2.0
scikit-learn==0.17.1
numpy==1.11.0
scipy==0.17.1
